+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * update Xcode project
 * remove shake-to-refresh
 * add Pierpont hours and verify other buildings' hours
+* support "Closed Soon" status and reorder statuses sanely
 * UI improvements:
     * redesign information display
     * new nav bar tint and table view section header colors